Fix double-logging in systemd
I am a fool. Of course sending logs to stdout/stderr would result in
them being logged by systemd. This begs the question, why was this set
up to log to systemd but NOT the streams in the first place?

Regardless, with this I hope I will never have to think about logging
again.